Well I'll be damned. Windows doesn't even like it if you put an int into a float ...

+++ src/uiTools/uipluginsel.cc
@@ -311,7 +311,7 @@ void reDrawnCB( CallBacker* )
     if ( parwdth < 1 || itmwdth < 1 )
 	{ pErrMsg("Huh"); return; }
 
-    const float diff = parwdth - itmwdth;
+    const int diff = parwdth - itmwdth;
     pmitem_->setPos( diff * .25f, 0.f );
 }
